requirements.txt
fastapi
uvicorn
requests
python-dotenv
genai_client
